requirements.txt
django
djangorestframework-simplejwt
django-cors-headers
pillow
psycopg2